move drawFont to rank
