+++ README.md
@@ -124,7 +124,7 @@ https://user-images.githubusercontent.com/32998901/127336796-bf4badcf-eaac-4b37-
 
 ## Codebase
 * `ansi` converts an image to ANSI art: colored text or color blocks (the latter can produce good image for pixel-style images).
-* `ascii` converts an image to ascii art. It utilize the rank generated by `rank`.
+* `ascii` converts an image to ascii art. It utilizes the rank generated by `rank`.
 * `rank` is used to rank the intensities of ascii symbols so that they can represent different grayscale values. Generally it is font-specific. My console uses monaco font, so I take it as an example. You are encouraged 
 to try other fonts by downloading the ttf file to directory `font`, and change the line 56 to your font file. Currently it is hardcoded, but can be set as cli arguments. Forgive me for not doing so because of lazyness.
 * `decode` extracts frames from gif/videos. Currently only gif is supported.
